# Add sumKernel4 with some PTX optimizations based on sumKernel3