From the NCU profiling result, the sumKernel3 is approaching the SOL bandwidth (~96.6% bandwidth throughput). Looking into the kernel code for a while, I thought that there are two potential minor optimizations we can make:

1. Bypassing L1 Cache: We only access the data once in the sum workload (the 0% L1 hit rate in the NCU report also proves this), so there is no need to allocate the L1 cache for these data when loading them from global memory. Therefore, we can use [`ld.global`](https://docs.nvidia.com/cuda/parallel-thread-execution/#data-movement-and-conversion-instructions-ld) with the [`.cg`](http://sw-mobile-docs/CUDA/GPGPU/parallel-thread-execution/#cache-operators) cache policy to bypass the L1 cache.
2. Warp-Wise Reduction: Utilize warp-wise reduction PTX code ([`redux.sync.add`](https://docs.nvidia.com/cuda/parallel-thread-execution/#parallel-synchronization-and-communication-instructions-redux-sync)) for the warp-wise summing.

After applying the above optimizations, I got a  ~0.015% speedup on H100 80GiB. Negligible, but we are closer to the SOL. 😅

```
$ make sum && out/sum
nvcc -std=c++17 -O3 -DNDEBUG -w --expt-relaxed-constexpr --expt-extended-lambda --use_fast_math -Xcompiler=-fPIE -Xcompiler=-Wno-psabi -Xcompiler=-fno-strict-aliasing -arch=sm_90a -lcublas -lcuda -lineinfo sum.cu -o out/sum
<---------------| kernel1 |--------------->
Bandwidth: 3236.51 GB/s
Time taken: 1.32704 ms
Sum: 1609549285

<---------------| kernel2 |--------------->
Bandwidth: 3241.19 GB/s
Time taken: 1.32512 ms
Sum: 1609549285

<---------------| kernel3 |--------------->
Bandwidth: 3241.26 GB/s
Time taken: 1.32509 ms
Sum: 1609549285

<---------------| kernel4 |--------------->
Bandwidth: 3241.8 GB/s
Time taken: 1.32487 ms
Sum: 1609549285

<---------------| cub |--------------->
Bandwidth: 3195.38 GB/s
Time taken: 1.34412 ms
Sum: 1609549285
```